+++ src/Instagram.php
@@ -1662,7 +1662,7 @@ class Instagram implements ExperimentsInterface
             } else {
                 $this->settings->set('phone_id', $this->settings->get('device_id'));
             }
-            $this->settings->set('uuid', Signatures::generateUUID());
+            $this->settings->set('uuid', Signatures::generateUUID(true, true));
 
             if ($loginType === 'facebook') {
                 $this->settings->set('fb_access_token', $password);